requirements.txt
absl-py==0.8.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
future==0.18.2
google-auth==1.8.1
google-auth-oauthlib==0.4.1
grpcio==1.25.0
idna==2.8
imageio==2.6.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.16.4
oauthlib==3.1.0
pandas==0.25.3
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.5
python-dateutil==2.8.1
pytorch-lightning==0.5.3.2
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.3.3
six==1.13.0
tensorboard==2.1.0
test-tube==0.7.5
torch==1.3.1
torchvision==0.4.2
tqdm==4.35.0
urllib3==1.25.7
Werkzeug==0.16.0
